fixed bug where commit hash would not be shown correctly
Before this was a library project, the value in the strings.xml file would be updated, the project would be built and everything worked correctly. Now that this is a library project, the strings from here are not compiled until the final complication of the project which is using this, so the ant build script we used before will no longer work. I modified it by creating a ~hash.xml file with the correct value which overrides the value in strings.xml. This file is ignored by git.